Move verification to a library function for hubtools
