+++ Cargo.toml
@@ -17,7 +17,9 @@ colored = { version = "2.0", default-features = false }
 crc-any = { version = "2.3.5", default-features = false }
 ecdsa = { version = "0.16.0", default-features = false, features = ["pkcs8", "der"] }
 elliptic-curve = { version = "0.13.1", default-features = false }
+env_logger = { version = "0.10", default-features = false, features = ["auto-color"] }
 hex = { version = "0.4.3", default-features = false, features = ["std"] }
+log = { version = "0.4", default-features = false }
 num-derive = { version = "0.3.0", default-features = false, features = ["full-syntax"] }
 num-traits = { version = "0.2.14", default-features = false }
 p256 = { version = "0.13.0", default-features = false, features = ["ecdsa", "pkcs8"] }
