+++ lpc55_sign/Cargo.toml
@@ -13,7 +13,9 @@ colored.workspace = true
 crc-any.workspace = true
 ecdsa.workspace = true
 elliptic-curve.workspace = true
+env_logger.workspace = true
 hex.workspace = true
+log.workspace = true
 lpc55_areas.workspace = true
 p256.workspace = true
 packed_struct.workspace = true
