+++ lpc55_sign/src/lib.rs
@@ -5,3 +5,4 @@
 pub mod crc_image;
 pub mod sign_ecc;
 pub mod signed_image;
+pub mod verify;
